# receive: Optimized receive; fixed long term overallocation