--- a/pkg/receive/writer.go
+++ b/pkg/receive/writer.go
@@ -64,8 +64,8 @@ func (r *Writer) Write(ctx context.Context, tenantID string, wreq *prompb.WriteR
 	for _, t := range wreq.Timeseries {
 		lset := labelpb.LabelsToPromLabels(t.Labels)
 
-		// Append as many valid samples as possible, but keep track of the errors.
 		var ref uint64
+		// Append as many valid samples as possible, but keep track of the errors.
 		for i, s := range t.Samples {
 			if i == 0 {
 				ref, err = app.Add(lset, s.Timestamp, s.Value)