--- a/pkg/store/labelpb/label.go
+++ b/pkg/store/labelpb/label.go
@@ -14,6 +14,7 @@ import (
 	"strings"
 	"unsafe"
 
+	"github.com/cespare/xxhash/v2"
 	"github.com/pkg/errors"
 	"github.com/prometheus/prometheus/pkg/labels"
 )
@@ -294,4 +295,33 @@ func DeepCopy(lbls []ZLabel) []ZLabel {
 		ret[i].Value = string(noAllocBytes(lbls[i].Value))
 	}
 	return ret
+}
+
+var sep = []byte{'\xff'}
+
+// HashWithPrefix returns a hash for the given prefix and labels.
+func HashWithPrefix(prefix string, lbls []Label) uint64 {
+	// Use xxhash.Sum64(b) for fast path as it's faster.
+	b := make([]byte, 0, 1024)
+	b = append(b, prefix...)
+
+	for i, v := range lbls {
+		if len(b)+len(v.Name)+len(v.Value)+2 >= cap(b) {
+			// If labels entry is 1KB allocate do not allocate whole entry.
+			h := xxhash.New()
+			_, _ = h.Write(b)
+			for _, v := range lbls[i:] {
+				_, _ = h.WriteString(v.Name)
+				_, _ = h.Write(sep)
+				_, _ = h.WriteString(v.Value)
+				_, _ = h.Write(sep)
+			}
+			return h.Sum64()
+		}
+		b = append(b, v.Name...)
+		b = append(b, sep[0])
+		b = append(b, v.Value...)
+		b = append(b, sep[0])
+	}
+	return xxhash.Sum64(b)
 }