… pkg/store/labelpb/label_test.go
… pkg/store/storepb/prompb/types.pb.go