@bwplotka 2020-10-17
Probably we can find more optimization, so far I can see this (inuse):
![image](https://user-images.githubusercontent.com/6950331/96352336-51f67580-10ba-11eb-8eed-4f6d3c402702.png)

(allocs):
![image](https://user-images.githubusercontent.com/6950331/96352371-7ce0c980-10ba-11eb-8221-5c6a6be482b9.png)